Change package price data type

+++ migrations/03-create-package.js
@@ -14,7 +14,7 @@ module.exports = {
       },
       price: {
         allowNull: false,
-        type: Sequelize.INTEGER
+        type: Sequelize.DECIMAL(15, 2)
       },
       createdAt: {
         allowNull: false,
